Renames, clear token, show signin options based on env
